feat(TaskModal.svelte): add support for running calculations on data source refinement
This commit adds the ability to run calculations on data source refinement in the TaskModal component. The runCalculation function was added to the component to handle the calculation submission. The function is a copy from the DataSource.svelte component. The function submits the calculation to the API using the setCalculation function from the api service. If there is only one engine available, the calculation is submitted immediately. If there are multiple engines available, the function sets the fragment to the edit-engine element for the current data source.